requirements.txt
chardet==5.2.0
gensim==4.3.2
itemadapter==0.7.0
jieba==0.42.1
matplotlib==3.5.1
numpy==1.24.4
pandas-2.2.2
PyMySQL==1.0.2
Scrapy==2.7.1
thulac==0.2.2
wordcloud==1.8.1
